Committer: Sarah Wooders <sarahwooders@gmail.com>

+++ skylark/replicate/replicator_client.py
@@ -1,8 +1,11 @@
+import atexit
+from contextlib import closing
 from datetime import datetime
 import itertools
 import json
 from logging import log
 from functools import partial
+from re import T
 import time
 from typing import Dict, List, Optional, Tuple
 
@@ -16,9 +19,10 @@ from skylark.benchmark.utils import refresh_instance_list
 from skylark.compute.aws.aws_cloud_provider import AWSCloudProvider
 from skylark.compute.gcp.gcp_cloud_provider import GCPCloudProvider
 from skylark.compute.server import Server, ServerState
-from skylark.gateway.chunk import Chunk, ChunkRequest, ChunkRequestHop, ChunkState
+from skylark.chunk import Chunk, ChunkRequest, ChunkRequestHop, ChunkState
 from skylark.obj_store.s3_interface import S3Interface
 from skylark.replicate.replication_plan import ReplicationJob, ReplicationTopology
+from skylark.replicate.replicator_client_dashboard import ReplicatorClientDashboard
 from skylark.utils.utils import PathLike, Timer, do_parallel, wait_for
 
 
@@ -77,10 +81,9 @@ class ReplicatorClient:
                     "instance_type": self.aws_instance_class,
                     "state": [ServerState.PENDING, ServerState.RUNNING],
                 }
-                with Timer("Refresh AWS instances"):
-                    current_aws_instances = refresh_instance_list(
-                        self.aws, set([r.split(":")[1] for r in aws_regions_to_provision]), aws_instance_filter
-                    )
+                current_aws_instances = refresh_instance_list(
+                    self.aws, set([r.split(":")[1] for r in aws_regions_to_provision]), aws_instance_filter
+                )
                 for r, ilist in current_aws_instances.items():
                     for i in ilist:
                         if f"aws:{r}" in aws_regions_to_provision:
@@ -104,7 +107,7 @@ class ReplicatorClient:
             else:
                 current_gcp_instances = {}
 
-        with Timer("Provision gateways"):
+        with Timer("Provisioning instances and waiting to boot"):
             # provision instances
             def provision_gateway_instance(region: str) -> Server:
                 provider, subregion = region.split(":")
@@ -119,12 +122,7 @@ class ReplicatorClient:
                     raise NotImplementedError(f"Unknown provider {provider}")
                 return server
 
-            results = do_parallel(
-                provision_gateway_instance,
-                list(aws_regions_to_provision + gcp_regions_to_provision),
-                progress_bar=True,
-                desc="Provisioning gateways",
-            )
+            results = do_parallel(provision_gateway_instance, list(aws_regions_to_provision + gcp_regions_to_provision))
             instances_by_region = {
                 r: [instance for instance_region, instance in results if instance_region == r] for r in set(regions_to_provision)
             }
@@ -155,21 +153,19 @@ class ReplicatorClient:
             bound_paths.append([instances_by_region[r].pop() for r in path])
         self.bound_paths = bound_paths
 
-        with Timer("Configure gateways"):
+        with Timer("Install gateway package on instances"):
             # start gateways
             do_parallel(
                 lambda s: s.start_gateway(
                     gateway_docker_image=self.gateway_docker_image, num_outgoing_connections=num_outgoing_connections
                 ),
                 list(itertools.chain(*self.bound_paths)),
-                progress_bar=True,
-                desc="Starting up gateways",
             )
 
     def deprovision_gateways(self):
         def deprovision_gateway_instance(server: Server):
-            logger.warning(f"Deprovisioning gateway {server.instance_name()}")
-            server.terminate_instance()
+            if server.instance_state() == ServerState.RUNNING:
+                server.terminate_instance()
 
         instances = [instance for path in self.bound_paths for instance in path]
         do_parallel(deprovision_gateway_instance, instances, n=len(instances))
@@ -222,15 +218,13 @@ class ReplicatorClient:
         assert len(chunk_batches) == n_src_instances, f"{len(chunk_batches)} batches, expected {n_src_instances}"
 
         # make list of ChunkRequests
-        chunk_requests_sharded: Dict[Server, List[ChunkRequest]] = {}
-        with tqdm(total=len(chunks), desc="Building chunk requests") as pbar:
-            for batch, path in zip(chunk_batches, self.bound_paths):
-                chunk_requests_sharded[path[0]] = []
+        chunk_requests_sharded: Dict[int, List[ChunkRequest]] = {}
+        with Timer("Building chunk requests"):
+            for batch, (path_idx, path_instances) in zip(chunk_batches, enumerate(self.bound_paths)):
+                chunk_requests_sharded[path_idx] = []
                 for chunk in batch:
-                    # make ChunkRequestHop list
                     cr_path = []
-                    for hop_idx, hop_instance in enumerate(path):
-                        # todo support object stores
+                    for hop_idx, hop_instance in enumerate(path_instances):
                         # TODO: cleanup...
                         src_object_store_region = None
                         src_object_store_bucket = None
@@ -250,8 +244,7 @@ class ReplicatorClient:
                                 dst_object_store_bucket = job.dest_bucket
                             else: # destination gateway
                                 location = "save_local"
-                        else:  # intermediate gateway
-                            location = "relay"
+                           location = "relay"
                         cr_path.append(
                             ChunkRequestHop(
                                 hop_cloud_region=hop_instance.region_tag,
@@ -264,15 +257,22 @@ class ReplicatorClient:
  
                             )
                         )
-                    chunk_requests_sharded[path[0]].append(ChunkRequest(chunk, cr_path))
-                    pbar.update()
+                    chunk_requests_sharded[path_idx].append(ChunkRequest(chunk, cr_path))
 
-        # send chunk requests to source gateway
-        for instance, chunk_requests in chunk_requests_sharded.items():
-            logger.debug(f"Sending {len(chunk_requests)} chunk requests to {instance.public_ip()}")
-            reply = requests.post(f"http://{instance.public_ip()}:8080/api/v1/chunk_requests", json=[cr.as_dict() for cr in chunk_requests])
-            if reply.status_code != 200:
-                raise Exception(f"Failed to send chunk requests to gateway instance {instance.instance_name()}: {reply.text}")
+        # send chunk requests to start gateways in parallel
+        with Timer("Dispatch chunk requests"):
+
+            def send_chunk_requests(args: Tuple[Server, List[ChunkRequest]]):
+                hop_instance, chunk_requests = args
+                logger.debug(f"Sending {len(chunk_requests)} chunk requests to {hop_instance.public_ip()}")
+                reply = requests.post(
+                    f"http://{hop_instance.public_ip()}:8080/api/v1/chunk_requests", json=[cr.as_dict() for cr in chunk_requests]
+                )
+                if reply.status_code != 200:
+                    raise Exception(f"Failed to send chunk requests to gateway instance {hop_instance.instance_name()}: {reply.text}")
+
+            start_instances = [(path[0], chunk_requests_sharded[path_idx]) for path_idx, path in enumerate(self.bound_paths)]
+            do_parallel(send_chunk_requests, start_instances, n=-1)
 
         return [cr for crlist in chunk_requests_sharded.values() for cr in crlist]
 
@@ -293,49 +293,75 @@ class ReplicatorClient:
         df = pd.DataFrame(chunk_logs)
         return df
 
-    def monitor_transfer(self, crs: List[ChunkRequest], completed_state=ChunkState.upload_complete) -> Dict:
+    def monitor_transfer(
+        self,
+        crs: List[ChunkRequest],
+        completed_state=ChunkState.upload_complete,
+        show_pbar=True,
+        log_interval_s: Optional[float] = None,
+        serve_web_dashboard=True,
+        dash_host="0.0.0.0",
+        dash_port="8080",
+        time_limit_seconds: Optional[float] = None,
+    ) -> Dict:
         total_bytes = sum([cr.chunk.chunk_length_bytes for cr in crs])
-        with tqdm(total=total_bytes * 8, desc="Replication", unit="bit", unit_scale=True, unit_divisor=KB) as pbar:
-            while True:
-                log_df = self.get_chunk_status_log_df()
-                last_log_df = (
-                    log_df.groupby(["chunk_id"])
-                    .apply(lambda df: df.sort_values(["path_idx", "hop_idx", "time"], ascending=False).head(1))
-                    .reset_index(drop=True)
-                )
-                is_complete_fn = (
-                    lambda row: row["state"] >= completed_state
-                    and row["path_idx"] == len(self.bound_paths) - 1
-                    and row["hop_idx"] == len(self.bound_paths[row["path_idx"]]) - 1
-                )
-                completed_chunk_ids = last_log_df[last_log_df.apply(is_complete_fn, axis=1)].chunk_id.values
-                completed_bytes = sum([cr.chunk.chunk_length_bytes for cr in crs if cr.chunk.chunk_id in completed_chunk_ids])
+        if serve_web_dashboard:
+            dash = ReplicatorClientDashboard(dash_host, dash_port)
+            dash.start()
+            atexit.register(dash.shutdown)
+            logger.info(f"Web dashboard running at {dash.dashboard_url}")
+        last_log = None
+        with Timer() as t:
+            with tqdm(
+                total=total_bytes * 8, desc="Replication", unit="bit", unit_scale=True, unit_divisor=KB, disable=not show_pbar
+            ) as pbar:
+                while True:
+                    log_df = self.get_chunk_status_log_df()
+                    if serve_web_dashboard:
+                        dash.update_status_df(log_df)
 
-                # update progress bar
-                pbar.update(completed_bytes * 8 - pbar.n)
-                total_runtime_s = (log_df.time.max() - log_df.time.min()).total_seconds()
-                throughput_gbits = completed_bytes * 8 / GB / total_runtime_s
-                pbar.set_description(f"Replication: average {throughput_gbits:.2f}Gbit/s")
+                    # count completed bytes
+                    last_log_df = (
+                        log_df.groupby(["chunk_id"])
+                        .apply(lambda df: df.sort_values(["path_idx", "hop_idx", "time"], ascending=False).head(1))
+                        .reset_index(drop=True)
+                    )
+                    is_complete_fn = (
+                        lambda row: row["state"] >= completed_state and row["hop_idx"] == len(self.bound_paths[row["path_idx"]]) - 1
+                    )
+                    completed_chunk_ids = last_log_df[last_log_df.apply(is_complete_fn, axis=1)].chunk_id.values
+                    completed_bytes = sum([cr.chunk.chunk_length_bytes for cr in crs if cr.chunk.chunk_id in completed_chunk_ids])
 
-                if len(completed_chunk_ids) == len(crs):
-                    return dict(completed_chunk_ids=completed_chunk_ids, total_runtime_s=total_runtime_s, throughput_gbits=throughput_gbits)
-                else:
-                    tqdm.write("\n\n")
-                    # print each path + hop (with region_tag) along with incomplete chunks and their status
-                    for path_idx, path in enumerate(self.bound_paths):
-                        for hop_idx, hop_instance in enumerate(path):
-                            last_log_df = (
-                                log_df[
-                                    (log_df.path_idx == path_idx)
-                                    & (log_df.hop_idx == hop_idx)
-                                    & (~log_df.chunk_id.isin(completed_chunk_ids))
-                                ]
-                                .groupby(["chunk_id"])
-                                .apply(lambda df: df.sort_values(["path_idx", "hop_idx", "time"], ascending=False).head(1))
-                                .reset_index(drop=True)
+                    # update progress bar
+                    pbar.update(completed_bytes * 8 - pbar.n)
+                    total_runtime_s = (log_df.time.max() - log_df.time.min()).total_seconds()
+                    throughput_gbits = completed_bytes * 8 / GB / total_runtime_s if total_runtime_s > 0 else 0.0
+                    pbar.set_description(f"Replication: average {throughput_gbits:.2f}Gbit/s")
+
+                    if len(completed_chunk_ids) == len(crs) or time_limit_seconds is not None and t.elapsed > time_limit_seconds:
+                        if serve_web_dashboard:
+                            dash.shutdown()
+                        return dict(
+                            completed_chunk_ids=completed_chunk_ids,
+                            total_runtime_s=total_runtime_s,
+                            throughput_gbits=throughput_gbits,
+                            monitor_status="completed" if len(completed_chunk_ids) == len(crs) else "timed_out",
+                        )
+                    else:
+                        current_time = datetime.now()
+                        if log_interval_s and (not last_log or (current_time - last_log).seconds > float(log_interval_s)):
+                            last_log = current_time
+                            gbits_remaining = (total_bytes - completed_bytes) * 8 / GB
+                            eta = int(gbits_remaining / throughput_gbits) if throughput_gbits > 0 else None
+                            logger.debug(
+                                f"{len(completed_chunk_ids)}/{len(crs)} chunks done ({completed_bytes / GB:.2f} / {total_bytes / GB:.2f}GB, {throughput_gbits:.2f}Gbit/s, ETA={str(eta) + 's' if eta is not None else 'unknown'})"
                             )
-                            for state, match_df in last_log_df.groupby("state"):
-                                tqdm.write(
-                                    f"[({path_idx}, {hop_idx}):{hop_instance.region_tag}] {state.name}: {', '.join(str(v) for v in match_df.chunk_id.values)}"
-                                )
-                    time.sleep(0.5)
+                        elif t.elapsed > 20 and completed_bytes == 0:
+                            logger.error(f"No chunks completed after {int(t.elapsed)}s! There is probably a bug, check logs. Exiting...")
+                            return dict(
+                                completed_chunk_ids=completed_chunk_ids,
+                                total_runtime_s=total_runtime_s,
+                                throughput_gbits=throughput_gbits,
+                                monitor_status="timed_out",
+                            )
+                        time.sleep(0.25)
